+++ rocket-okapi/src/request/from_segments_impls.rs
@@ -1,6 +1,6 @@
 use super::OpenApiFromSegments;
 use crate::gen::OpenApiGenerator;
-use okapi::openapi3::{Parameter, ParameterValue};
+use okapi::openapi3::{Object, Parameter, ParameterValue};
 use schemars::JsonSchema;
 
 type Result = crate::Result<Parameter>;
@@ -26,7 +26,7 @@ where
                 example: None,
                 examples: None,
             },
-            extensions: okapi::Map::default(),
+            extensions: Object::default(),
         })
     }
 }